Move all supertux/ files
To the dir supertux/old/.
SuperTux Editor code will be under the GPLv3+,
since I want to use some of the code which already exists.
