+++ supertux/TODO.md
@@ -1,0 +1,3 @@
+# SuperTux Editor TODO
+
+*There's nothing here*
