+++ supertux/old/gameobj.py
