+++ supertux/old/gameobj_factor.py
