+++ supertux/old/gameobj_props_change_command.py
