+++ supertux/old/sector.py
