+++ supertux/old/sprite.py
