+++ supertux/old/util.py
