+++ supertux/old/worldmap_object.py
